ProjectSelect is its own component
